split aiderCommand modules
