+++ denops/aider/mockAiderCommand.ts
@@ -1,5 +1,4 @@
 import { emit } from "https://deno.land/x/denops_std@v6.4.0/autocmd/mod.ts";
-// deno-lint-ignore-file require-await
 import * as fn from "https://deno.land/x/denops_std@v6.4.0/function/mod.ts";
 import type { Denops } from "https://deno.land/x/denops_std@v6.4.0/mod.ts";
 import type { AiderCommands } from "./aiderCommand.ts";
@@ -35,6 +34,7 @@ export const commands: AiderCommands = {
     await denops.cmd("bd!");
   },
 
+  // deno-lint-ignore require-await
   checkIfAiderBuffer: async (_: Denops, bufnr: number): Promise<boolean> => {
     return bufnr === mockAiderBufnr;
   },
